shoppingmall/src/main/java/kr/megaptera/shoppingMall/services/ReviewService.java
package kr.megaptera.shoppingMall.services;

import kr.megaptera.shoppingMall.dtos.ReviewDto;
import kr.megaptera.shoppingMall.models.Review;
import kr.megaptera.shoppingMall.repositoies.ReviewRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class ReviewService {
  private ReviewRepository reviewRepository;

  public ReviewService(ReviewRepository reviewRepository) {
    this.reviewRepository = reviewRepository;
  }

  public List<Review> listByProductId(Long productId) {
    return reviewRepository.findAllByProductId(productId);
  }
}
